Adding Exceptions, ExceptionLimiter class, Notifier class, and PrinterQueue class
Exceptions.rb contains a custom exception class (one which can accept an object as well as a message) and a list of all the exceptions used in the application.

ExceptionLimiter.rb - maintains a list of currently, or recently raised exceptions

Notifier.rb - handles converting exceptions to emails and sends them. Uses the exception's class name as the email sender method. Uses ExceptionLimiter

PrinterQueue.rb - handles communication with the printer and parses the queue's status.